audio_sim/requirements.txt
pandas
numpy>=1.21.0,<2.0.0
sentence-transformers
faiss-gpu
torch
torchvision
torchaudio
librosa
fastapi
uvicorn
